requirements.txt
ultralytics
streamlit
opencv-python-headless
Pillow
numpy
torch
torchvision